+++ AdventOfCode2023/ISolution.cs
@@ -4,7 +4,7 @@ internal interface ISolution
 {
     int Day { get; }
 
-    int SolveFirstPuzzle(IReadOnlyList<string> input);
+    long SolveFirstPuzzle(IReadOnlyList<string> input);
 
-    int SolveSecondPuzzle(IReadOnlyList<string> input);
+    long SolveSecondPuzzle(IReadOnlyList<string> input);
 }
